@@ -5,9 +5,12 @@ source 'https://rubygems.org'
 # development dependencies will be added by default to the :development group.
 gemspec
 
-gem 'rails', '~> 4.2.7'
-gem 'bundler', '~> 1.14'
+gem 'rails', '~> 5.0'
+gem 'bundler', '~> 2.1.4'
 
+group :test do
+  gem 'reform-rails'
+end
 
 # Declare any dependencies that are still in development here instead of in
 # your gemspec. These might include edge Rails or gems from your path or
