@@ -37,13 +37,21 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency "rspec-rails", "~> 3.0"
   spec.add_development_dependency 'sqlite3'
   spec.add_development_dependency "factory_bot_rails"
-  spec.add_development_dependency "shoulda-matchers"
+  spec.add_development_dependency 'shoulda-matchers', '~> 4.0.0.rc1'
+  spec.add_development_dependency "capybara"
+  spec.add_development_dependency "selenium-webdriver"
+  spec.add_development_dependency "recaptcha"
+  spec.add_development_dependency "puma", ">= 4.3.8"
+  spec.add_development_dependency 'jquery-rails'
+  spec.add_development_dependency 'byebug'
 
-
-  spec.add_dependency 'rails', '~> 4.2.0'
-  spec.add_dependency 'reform', '~> 1.2'
-  spec.add_dependency 'wicked', '~> 1.1'
-  spec.add_dependency 'decent_exposure', '~> 3.0'
+  spec.add_dependency 'rails', '~> 5.0'
+  spec.add_dependency 'reform', '2.2.4'
+  spec.add_dependency 'execjs', '2.7.0'
+  spec.add_dependency 'wicked'
+  spec.add_dependency 'decent_exposure'
   spec.add_dependency 'haml'
+  spec.add_dependency 'nfg_ui', '~> 0.13.0'
 
+  spec.add_dependency 'simple_form'
 end
