@@ -1,4 +1,4 @@
-class CreateAdmin < ActiveRecord::Migration
+class CreateAdmin < ActiveRecord::Migration[5.0]
   def change
     create_table :admins do |t|
       t.string :first_name
