@@ -1,4 +1,4 @@
-class CreateOnboardingRelatedObject < ActiveRecord::Migration
+class CreateOnboardingRelatedObject < ActiveRecord::Migration[5.0]
   def change
     create_table "onboarding_related_objects", force: :cascade do |t|
       t.string  "name",                  :limit=>255
