@@ -1,7 +1,11 @@
-module NfgOnboarder::OnboardableObject
-  extend ActiveSupport::Concern
+# frozen_string_literal: true
 
-  included do
-    has_many :related_objects, as: :target, class_name: 'Onboarding::RelatedObject', dependent: :destroy
+module NfgOnboarder
+  module OnboardableObject
+    extend ActiveSupport::Concern
+
+    included do
+      has_many :related_objects, as: :target, class_name: 'Onboarding::RelatedObject', dependent: :destroy
+    end
   end
 end
