@@ -1,3 +1,3 @@
 class NfgOnboarder::InformationalForm < Reform::Form
-
+  model NfgOnboarder::InformationalForm
 end
