@@ -17,15 +17,15 @@ module NfgOnboarder
       end
 
       def add_steps_to_controller
-        inject_into_file controller, after: "# steps list\n" do <<-STRING
-  steps #{ steps_to_symbols }
+        inject_into_file controller, after: "def self.step_list\n" do <<-STRING
+    %i[#{ steps.join(" ") }]
         STRING
         end
       end
 
       def add_on_valid_methods_to_controller
         steps.each do |step_name|
-          inject_into_file controller, after: "# on valid steps\n" do <<-STRING
+          inject_into_file controller, before: "  # end on valid steps" do <<-STRING
   def #{ step_name }_on_valid_step
     # you can add logic here to perform actions once a step has completed successfully
   end
@@ -37,9 +37,9 @@ module NfgOnboarder
 
       def add_on_before_save_methods_to_controller
         steps.each do |step_name|
-          inject_into_file controller, after: "# on before save steps\n" do <<-STRING
+          inject_into_file controller, before: "  # end on before save steps" do <<-STRING
   def #{ step_name }_on_before_save
-    # you can add logic here to perform, such as appending data to the params, before the form is to be saved
+    # you can add logic here to perform an action, such as appending data to the params, before the form is to be saved
   end
 
           STRING
@@ -48,11 +48,11 @@ module NfgOnboarder
       end
 
       def add_get_form_target_entry
-        inject_into_file controller, after: "case step\n" do
+        inject_into_file controller, before: "    # catch all" do
           steps.inject("") do |str, step_name|
             str += <<-STRING
-        when :#{ step_name.underscore }
-          OpenStruct.new(name: '') # replace with your object that the step will update
+    when :#{ step_name.underscore }
+      OpenStruct.new(name: '') # replace with your object that the step will update
     STRING
           end
         end
@@ -88,15 +88,16 @@ module NfgOnboarder
       def add_entries_to_locals_file
         template "locales.rb", locales_file unless File.exists?(locales_file)
         if onboarding_group.present?
+          # add the language areas for each page
           steps.each do |step_name|
-            inject_into_file locales_file, after: "#{onboarder_name.underscore}:\n" do <<-STRING
+            inject_into_file locales_file, before: "        step_navigations:" do <<-STRING
         #{step_name.underscore}:
           header:
             message: 'Replace me at config/locales/view/onboarding/#{onboarder_name.underscore}'
             form: 'Replace me too'
             page:
           button:
-            submit: Next
+            <<: *default_buttons
           label:
           placeholder:
           hint:
@@ -105,16 +106,25 @@ module NfgOnboarder
           STRING
             end
           end
-        else
+
+          # add the steps to the navigation area
           steps.each do |step_name|
-            inject_into_file locales_file, after: "#{onboarder_name.underscore}:\n" do <<-STRING
+            inject_into_file locales_file, before: "# end of file" do <<-STRING
+          #{step_name.underscore}: #{step_name.humanize}
+          STRING
+            end
+          end
+        else
+          # add the language areas for each page
+          steps.each do |step_name|
+            inject_into_file locales_file, before: "      step_navigations:" do <<-STRING
       #{step_name.underscore}:
         header:
           message: 'Replace me at config/locales/view/onboarding/#{onboarder_name.underscore}'
           form: 'Replace me too'
           page:
         button:
-          submit: Next
+          <<: *default_buttons
         label:
         placeholder:
         hint:
@@ -123,12 +133,19 @@ module NfgOnboarder
           STRING
             end
           end
+          # add the steps to the navigation area
+          steps.each do |step_name|
+            inject_into_file locales_file, before: "# end of file" do <<-STRING
+        #{step_name.underscore}: #{step_name.humanize}
+          STRING
+            end
+          end
         end
       end
 
       def add_step_navigations
         steps.each do |step_name|
-          inject_into_file locales_file, after: "step_navigations:\n" do
+          inject_into_file locales_file, before: "# end of file" do
         "    #{step_name.underscore}:\n"
           end
         end
@@ -211,10 +228,6 @@ module NfgOnboarder
 
       def onboarder_and_group_name_split
         onboarder_and_group_name.split("::")
-      end
-
-      def steps_to_symbols
-        steps.map { |step_name| ":#{ step_name.underscore }" }.join(", ")
       end
 
       # def add_step_to_controller
