@@ -1,14 +1,30 @@
 class Onboarding::<%= combined_onboarder_name %>Controller < <%= inherited_controller_name %>Controller
 
   # steps list
+  # we do this so we can access the list of steps from outside the onboarder
+  def self.step_list
+  end
+
+  # if a step is included in exit_without_save_steps, it will exit onboarder without saving.
+  # it will also show 'exit' button instead of 'save and exit'
+  # otherwise it will save and exit
+
+  def exit_without_save_steps
+    []
+  end
+
+  # steps list
+  steps *step_list
 
   expose(:onboarding_group_steps) { [] } # this should be removed if you are using a group step controller as a parent of this controller
 
   private
 
-  # on before save steps
+  # begin on before save steps
+  # end on before save steps
 
-  # on valid steps
+  # begin on valid steps
+  # end on valid steps
 
   def can_view_step_without_onboarding_session
     return true if params[:id] == 'wicked_finish' # the onboarding session is typically completed prior to this step
@@ -19,9 +35,24 @@ class Onboarding::<%= combined_onboarder_name %>Controller < <%= inherited_contr
 
   def get_form_target
     case step
+    # catch all
     else
       OpenStruct.new(name: '')
     end
+  end
+
+  def get_onboarding_session
+    # we have to find the onboarding session first from the user session, if we can't find it then we need to look at the params
+    # if still can't find it then we create a new onboarding session
+    onboarding_sess = nil
+    onboarding_sess = ::Onboarding::Session.find_by(id: session[:onboarding_session_id]) if session[:onboarding_session_id]
+    onboarding_sess ||= new_onboarding_session
+    onboarding_sess.tap { |os| session[:onboarding_session_id] = os.id }
+  end
+
+  def fields_to_be_cleansed_from_form_params
+    # these are fields that we don't want to store in onboarder session
+    %w{}
   end
 
   def finish_wizard_path
@@ -29,24 +60,28 @@ class Onboarding::<%= combined_onboarder_name %>Controller < <%= inherited_contr
      # TODO add a path to where the user should go once they complete the onboarder
   end
 
-  def onboarder_name
-    "<%= onboarder_name.underscore %>"
+  def new_onboarding_session
+    ::Onboarding::Session.create(onboarding_session_parameters)
   end
 
-  def get_onboarding_session
-    # Use the following as an example of how an onboarding session would be either retrieved or instantiated
-    # We call new rather than create because we don't want the onboarding session
-    # to be saved if the user does not continue past the first step.
-    # onboarding_admin.onboarding_session_for(onboarder_name) || Onboarding::Session.new(onboarding_session_parameters)
+  def onboarder_name
+    "<%= onboarder_name.underscore %>_onboarder"
   end
 
   def onboarding_session_parameters
     {
       entity: nil,# supply the parent object to the onboarding session
       owner: onboarding_admin,
-      current_step: ,  #typically the first step
+      current_step: "TODO: Update to the step on which you want to create the onboarding session",  #typically the first step
       related_objects: {} ,# a hash containing the whatever object will be saved first, i.e. { project: get_project },
       name: onboarder_name
     }
+  end
+
+  def reset_onboarding_session
+    # when you need to clear the onboarding session (so the user can start a new onboarder)
+    # call this method
+    session[:onboarding_session_id] = nil
+    session[:onboarding_import_data_import_id] = nil
   end
 end
