@@ -1,4 +1,12 @@
-class Onboarding::CreateProject::ProjectDescriptionForm < Reform::Form
-  ## Add properties for your form below:
-  # property :my_property
+# frozen_string_literal: true
+
+module Onboarding
+  module CreateProject
+    class ProjectDescriptionForm < Onboarding::CreateProject::BaseForm
+      ## Add properties for your form below:
+      property :description
+
+      validates :description, presence: true
+    end
+  end
 end
