@@ -1,4 +1,12 @@
-class Onboarding::CreateProject::ProjectNameForm < Reform::Form
-  ## Add properties for your form below:
-  # property :my_property
+# frozen_string_literal: true
+
+module Onboarding
+  module CreateProject
+    class ProjectNameForm < Onboarding::CreateProject::BaseForm
+      ## Add properties for your form below:
+      property :name
+
+      validates :name, presence: true
+    end
+  end
 end
